extract_tv_ILV()

private function to extract the tv ILV matrices and transform them into ILV_tv dataframe for further processing This is a really silly way of doing it:

the original TV matrix is impossible to reconstruct from NBDA objs, since NBDA only saves values for when individuals are naive. This shouldn't matter for basic functionality, but if you have a use-case where you wanted to do counterfactuals etc, one needs that data.

this basically converts vector > matrix > vector, which will then be converted again into a matrix by the import_NBDA_STb function.. However it works, will come back later to make it more efficient.

Usage

extract_tv_ILV(nbda_object, ILV_type)

Arguments

nbda_object

NBDAdata object

ILV_type

character indicating types of ILVs: @asocILVdata = asocial, @intILVdata = social, and @multiILVdata = multiplicative (affects both social and asocial equally)

Value

List of vectors for each ILV