generate_STb_model_TADA()

Helper function that creates stan code for cTADA and dTADA type models.

Usage

generate_STb_model_TADA(
  STb_data,
  model_type = "full",
  intrinsic_rate = "constant",
  transmission_func = "standard",
  dTADA = F,
  veff_ID = c(),
  gq = TRUE,
  est_acqTime = FALSE,
  priors = list()
)

Arguments

STb_data

a list of formatted data returned from the STbayes_data() function

model_type

string specifying the model type: "asocial" or "full"

intrinsic_rate

string specifying whether intrinsic rate is static or can change over time ("constant", "weibull")

transmission_func

string specifying transmission function: "standard", "freqdep_f" or "freqdep_k" for complex contagion. Defaults to "standard". Complex contagion with multi-network model is not supported.

dTADA

boolean indicating whether dTADA should be used.

veff_ID

Parameters for which to estimate varying effects by individuals. Default is no varying effects.

gq

Boolean to indicate whether the generated quantities block is added (incl. ll for WAIC)

est_acqTime

Boolean to indicate whether gq block includes estimates for acquisition time. At the moment this uses 'one weird trick' to accomplish this and does not support estimates for non-integer learning times.

priors

named list with strings containing the prior for log baserate, s, f, k.

Value

A STAN model (character) that is customized to the input data.