import_NBDA_STb()

Create STbayes_data object from nbda object

Usage

import_NBDA_STb(
  nbda_object,
  network_names = c("default"),
  ILVi = NULL,
  ILVs = NULL,
  ILVm = NULL
)

Arguments

nbda_object

object of NBDAdata class

network_names

character vector of descriptive names of networks you're importing.

ILVi

Optional character vector of ILVS to be considered when estimating intrinsic rate. If not specified, taken from NBDA object.

ILVs

Optional character vector of ILVS to be considered when estimating social transmission rate. If not specified, taken from NBDA object.

ILVm

Optional character vector of ILVS to be considered in a multiplicative model. If not specified, taken from NBDA object.

Value

A list containing properly formatted data to run social transmission models.

Examples

nbda_object <- STbayes::tutorial4_1
data_list <- import_NBDA_STb(nbda_object, network_names = "assoc")